backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Trello-lite"
authors = [
    {name = "Oatmeal4Breakfast", email = "elvin.salcedo17@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pydantic[email]>=2.8.2,<3.0.0",
    "pyjwt>=2.10.1",
]
